py/data/ffmpeg-notes/pyproject.toml
[project]
name = 'ffmpeg-notes'
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["ffmpeg-python>=0.2.0"]

[tool.uv]
dev-dependencies = ["ruff", "pylint"]

[tool.ruff.lint]
extend-select = [
    "I",
    "W",
    "E",
    "D",
    "UP",
    "B",
    "SIM",
    "F",
    "S",
    "A",
    "C4",
    "ISC",
    "TCH",
    "PTH",
    "PL",
    "PERF",
    "FURB",
    # "RUF",
]

[tool.pyright]
typeCheckingMode = "basic"
